fix(theme): Wrap around `useThemeContext` to update components.
If you load the site with light theme enabled, the message components still load in dark theme (probably something on Docusaurus's side.) So, this is a thin wrapper that will cause elements to refresh to light theme once the site is loaded. (No performance issues so far.)

Also, the embed titles were somehow still white in light theme. No clue how that happens, but I included custom CSS to resolve that (for now).